userlandfs: Fix the BeOS frontend on 64-bit.

+++ src/add-ons/kernel/file_systems/userlandfs/server/beos/lock.h
@@ -24,7 +24,7 @@
 
 struct beos_lock {
 	sem_id		s;
-	long		c;
+	int32		c;
 };
 
 struct beos_mlock {
